:hammer_and_wrench: improve plugin handler
Split files, add Ng converters, use dumb integration test.
