+++ libs/httpext/src/lib/utils/is-promise.ts
@@ -1,0 +1,3 @@
+export function isPromise<T>(value: any): value is Promise<T> {
+  return typeof value.then === 'function';
+}
